# Add performance comparison against Scala namer.