Currently rsc does not have a type-checker that makes for a meaningful comparison against the Scala type-checker. Since Rsc currently do mostly name-resolution, a comparison against a similar phase in Scalac is useful.

|                    | Cold                 | Hot                  |
|--------------------|----------------------|----------------------|
| RscTypecheck       | 476.843 ± 0.822 ms   | 33.685 ± 0.024 ms    |
| ScalacNamer 2.12.4 | 1843.751 ± 14.930 ms | 29.825 ± 0.090 ms    |
| ScalacNamer 2.11.11| 1423.331 ± 37.056 ms | 71.100 ± 0.318 ms    |


The new numbers are interesting: compared to the namer in 2.12.4, rsc is 10% slower, but about 2x faster than the 2.11.11 namer in "hot" state. Things look a lot better for cold performance, where rsc is about 3x faster than the Scala namer.