@retronym 2017-11-20
Is this really a fair comparison? I think it might be fairer to run `-Ystop-after:typer`, but use a modified `Typer` that discards the RHS of `DefDef` and `ValDef`.