@dragos 2017-11-20
That would be a useful data point, though it would not be much fairer. I couldn't find the code in Rsc that does any sort of type-checking of classes (parents, bound checks, etc), so even if you remove method and field bodies you'd still do a lot more work in the scalac type-checker. As far as I can tell, at this point the Rsc type-checker is a traverser that creates entries in the symbol table, but not much more than that.

Perhaps any kind of benchmarking is premature, but if there is one, I think the namer is probably closer to the truth than even the modified typer you suggested.