@retronym 2017-11-20
I'm really eager to learn from this experiment, but I'm happy to wait a little while :)

Here's a simple/cheap way to modify scalac to drop RHS-s (basically add a traversal phase post-parser that mutates `DefDef` and `ValDef`-s.

https://github.com/retronym/scalac-survival-guide/commit/cec68be7d8e5a7d96af68dd61c896629429dd855

But as you say, this probably isn't worth trying out just to get a apples/guava comparison in addition to the apples/oranges :)