@gkossakowski 2017-11-20
An alternative v0 step is to make a synthetic input where methods have RHS empty by being abstract. And keep the method signatures really simple to exercise two pieces: entering symbols, resolving names.

That's what I did in Kentucky Mule to get an idea what's the gap between what scalac's performance and what's the limit of computers speed: https://github.com/gkossakowski/kentuckymule#is-kentucky-mule-really-fast
I didn't have even an intuition how quickly name resolution can happen. Having realized name resolution can be much faster than I expected, I booked that as a win and had the confidence to investigate further.

The larger point I would like to make is the early numbers are not going to be great for apple-to-apple comparison but just as a source of negative validation. If early on, without many of the features implemented, your numbers are not looking extremely well in comparison to `scalac` then there's nothing else to investigate in that direction. That's how I thought of the numbers I got while working on Kentucky Mule.