@xeno-by 2017-11-24
Superseded by https://github.com/twitter/reasonable-scala/pull/27.